tmp175: Decrease "master" member variable of tmp175_data
The master member variable doesn't need 16 bits.  Change it to
uint8_t.  This decrease the total data space from 317 bytes to 313
bytes.

+++ src/tmp175.h
@@ -18,7 +18,7 @@
 #define REG_THIGH  0x03
 
 struct tmp175_data {
-        int  master;
+        uint8_t master;
         uint8_t addr;
         uint8_t data[2];
         uint8_t error;
